chore(registration-react): split index to use App component
